refact: rename package core.internal.types to core.internal.grammar.raw
1. renames package org.eclipse.tm4e.core.internal.types to
org.eclipse.tm4e.core.internal.grammar.raw
2. moves related classes from org.eclipse.tm4e.core.internal.grammar to
org.eclipse.tm4e.core.internal.grammar.raw
