+++ org.eclipse.tm4e.core/src/main/java/org/eclipse/tm4e/core/internal/grammar/raw/package-info.java
@@ -4,6 +4,6 @@
  * github.com/microsoft/vscode-textmate/blob/main/src/rawGrammar.ts</a>
  */
 @NonNullByDefault
-package org.eclipse.tm4e.core.internal.types;
+package org.eclipse.tm4e.core.internal.grammar.raw;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
