+++ org.eclipse.tm4e.core/src/main/java/org/eclipse/tm4e/core/internal/theme/raw/IRawThemeSetting.java
